fixed the typos and renamed the package
